refactor(lotto package): 패키지 및 변수명 변경, 불필요한 로직 제거
- 도메인 모델을 model 패키지 하위로 이동
- purchasePrice -> purchaseAmount로 통일
- Lotto 모델의 isWin 메소드 제거 및 WinPolicy 제거 (사용하지 않음)
- Money의 divide 예외 제거 (정상적인 프로그램 동작을 위해 0으로 나눌 때 0으로 리턴하도록 수정)
