+++ src/main/java/lotto/model/lotto/LottoNumberGeneratorStrategy.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.lotto;
 
 import java.util.List;
