+++ src/main/java/lotto/model/lotto/LottoNumbers.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.lotto;
 
 import java.util.ArrayList;
 import java.util.List;
