--- src/main/java/lotto/win/DefaultWinPolicy.java
@@ -1,9 +1,0 @@
-package lotto.win;
-
-public class DefaultWinPolicy implements WinPolicy {
-
-    @Override
-    public boolean isWin(int matchCount) {
-        return WinRanking.isWin(matchCount);
-    }
-}
