--- src/main/java/lotto/win/WinPolicy.java
@@ -1,5 +1,0 @@
-package lotto.win;
-
-public interface WinPolicy {
-    boolean isWin(int matchCount);
-}
