+++ src/test/java/lotto/LottoNumberGeneratorTest.java
@@ -2,6 +2,9 @@ package lotto;
 
 import java.util.HashSet;
 import java.util.List;
+import lotto.model.lotto.DefaultNumberGeneratorStrategy;
+import lotto.model.lotto.LottoNumber;
+import lotto.model.lotto.LottoNumberGenerator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
