+++ src/test/java/lotto/money/MoneyTest.java
@@ -1,8 +1,8 @@
 package lotto.money;
 
+import lotto.model.money.Money;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
@@ -16,15 +16,6 @@ class MoneyTest {
     void moneyDivide(double input, double expected) {
         double result = money.divide(Money.from(input));
         Assertions.assertThat(result).isEqualTo(expected);
-    }
-
-    @Test
-    @DisplayName("Money의 divide 메소드에서 money.value가 0인 경우 IllegalArgumentException을 던진다.")
-    void moneyDivideException() {
-        Money zero = Money.from(0);
-
-        Assertions.assertThatThrownBy(() -> money.divide(zero))
-                .isInstanceOf(IllegalArgumentException.class);
     }
 
     @ParameterizedTest
