+++ src/test/java/lotto/statistic/LottoStatisticTest.java
@@ -2,12 +2,14 @@ package lotto.statistic;
 
 import java.util.Arrays;
 import java.util.stream.Stream;
-import lotto.Lotto;
-import lotto.LottoNumbers;
-import lotto.Lottos;
+import lotto.model.lotto.Lotto;
+import lotto.model.lotto.LottoNumbers;
+import lotto.model.lotto.Lottos;
 import lotto.TestLottoNumberGeneratorStrategy;
-import lotto.money.Money;
-import lotto.win.WinRanking;
+import lotto.model.money.Money;
+import lotto.model.statistic.LottoStatistic;
+import lotto.model.statistic.LottoStatisticResults;
+import lotto.model.win.WinRanking;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
